Added admin checking
There is a file called "isadmin.jsp" in the includes folder that every
admin page must include.  This file will check the header for the "user"
data.  If it's not there, it will redirect to the log in page/index.  If
it's there but it's not an admin, it will redirect to the view/theme
page.  Otherwise, it will go to the admin pages.