Horray, it's on Amazon!
techtoberfest.elasticbeanstalk.com for those interested.

Note: there's something odd about the database on amazon.  It thinks
it's in Pacific time, despite me making sure the environment was based
in Virginia, not Oregon.  I adjusted one of the queries to take this
into account on the attendance page.  It will likely need to be changed
in the future.

Headerdump.jsp is a helper so Larry York and team can point the
SiteMinder header stuff at our site.