Mobile-friendly navigation headers
Or at least, more mobile-friendly.  The respond.min.js file adapts media
queries for older browsers, so it'll work on IE8, etc.  The navs use
Scripps Bootstrap's hidden-desktop/tablet/phone classes to show and hide
different sized information.  Probably not the most elegant way of doing
it, but it works.