View Session Schedule & Manage Session Schedule were throwing Null Pointer Exception. I added a try catch block to try to patch the problem until we can figure out what's wrong.
NOTE: somehow I added a 2nd Test Session 8, this caused the Session to show up as Null which was causing some problems, nevertheless the site should not crash because of one bad session. I added so try-catch blocks, but this should be fixed later on. The sit should not break from user entered data, even if that user is an admin.